twirp-app/pyproject.toml
[tool.poetry]
name = "twirp-app"
version = "0.1.0"
description = ""
authors = ["Nick Schofield <nschofield@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
twirp = "^0.0.3"
uvicorn = "^0.13.3"
ipython = "^7.19.0"
PyJWT = "^2.0.0"
python-dotenv = "^0.15.0"
requests = "^2.25.1"
SQLAlchemy = "^1.3.22"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
